Updating transaction queries
* Updated transactions queries per #14
* Upgraded Exposed libraries from 0.17.1 to 0.29.1
* Updated native queries to exposed queries

closes: #14

+++ service/src/main/kotlin/io/provenance/explorer/domain/extensions/Extenstions.kt
@@ -104,7 +104,7 @@ fun <T> List<T>.pageOfResults(page: Int, count: Int): List<T> {
 }
 
 // Total # of results divided by count per page
-fun Int.pageCountOfResults(count: Int): Int =
+fun Long.pageCountOfResults(count: Int): Int =
     if (this < count) 1
     else ceil(this.toDouble() / count).toInt()
